requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
requests==2.31.0
beautifulsoup4==4.12.2
gunicorn==21.2.0
psycopg2-binary==2.9.9
python-Levenshtein==0.23.0
